% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoVis.R
\name{attenceVis}
\alias{attenceVis}
\title{visualize status}
\usage{
attenceVis(df, ids = NULL)
}
\arguments{
\item{df}{dataframe about users' geo information, including id, account_id,
longitude, latitude and status field at least.}

\item{ids}{integer vector, indicate the users whose status will be change to
1 and their cirlces on map will be blue.}
}
\description{
visualize user's longitude and latitude on map
}
\details{
see users' status information on the map, the status field had already been
settle by SQL store procedure on the remote server, if \code{ids} were
provided, the circle represent ids will also been labeled as \code{status == 1}.
blue circles indicate \code{status == 1}, and red ones indicate opposite.
}

