% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{coocCreate}
\alias{coocCreate}
\title{make co-occurrence matrix}
\usage{
coocCreate(df, base = "item", method = "co_times")
}
\arguments{
\item{df}{dataframe about users , items and rating, if \code{method} is
\code{co_times} or \code{jaccard}, users and items columns are need only.}

\item{base}{\code{user} or \code{item}, difference collective filtering algorithm.}

\item{method}{\code{co_times}, \code{jaccard} or \code{rating_sim}}
}
\description{
make co-occurrence matrix based on users or items
}
\details{
\itemize{
 \item \code{co_times} method will simpley get the co-occurrence times as result matrix
cell value,
 \item \code{jaccard} go one more step based on \code{co_times} method, it will
caculate two items(or users) jaccard coefficient as matrix cell value,
 \item \code{rating_sim} use algorithm made by myself, it uses rating rating ratio
between two items(or users) rated by one user(or item) as matrix cell value.
}
}
