% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{jaccardCaculate}
\alias{jaccardCaculate}
\title{jaccard coefficient}
\usage{
jaccardCaculate(a, b)
}
\arguments{
\item{a, b}{same class vectors}
}
\description{
caculate jaccard coefficient between two sets
}
\details{
jaccard coefficient was caculated by the formula:
\deqn{\frac{a\bigcap b}{a\bigcup b}}
}

