% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{scaleSafe}
\alias{scaleSafe}
\title{caculate scale safely}
\usage{
scaleSafe(vec, center = TRUE, scale = TRUE)
}
\arguments{
\item{vec}{numeric vector}

\item{center, scale}{same as \code{scale} parameters}
}
\description{
scale data, almost equal with \code{scale}, but safely
}
\details{
when vec is all equal, \code{scale} will give all NAN results, which is not
acceptable for the fallowing caculation, \code{scaleSafe} will give all 0 vector
in this case.
}
\seealso{
scale
}

