% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceCaculate.R
\name{distanceCaculate}
\alias{distanceCaculate}
\title{caculate distance}
\usage{
distanceCaculate(lon_1, lat_1, lon_2, lat_2)
}
\arguments{
\item{lon_1, lon_2}{numeric, degree longitude}

\item{lat_1, lat_2}{numeric, degree latitude}
}
\details{
caculate the spherical distance between 2 point on earth, although
\code{geosphere::distHaversine} can caculate the spherical distance,
\code{distanceCaculate} has about 10X speed than it and less than 0.2% error.
}
\examples{
distanceCaculate(108.8821, 34.2341, 108.882, 34.2343)
}

