% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceCaculate.R
\name{distanceCaculate}
\alias{distanceCaculate}
\title{caculate distance}
\usage{
distanceCaculate(lon_1, lat_1, lon_2, lat_2, r = 6378137)
}
\arguments{
\item{lon_1, lon_2}{numeric, degree longitude}

\item{lat_1, lat_2}{numeric, degree latitude}

\item{r}{earth radius from \code{geosphere} package}
}
\description{
caculate the spherical distance between 2 point on earth
}
\details{
although \code{geosphere::distHaversine} can caculate the spherical distance,
\code{distanceCaculate} is about 20X-30X faster than that,
and only has error less than 0.2\%.
}
\examples{
distanceCaculate(108.8821, 34.2341, 108.882, 34.2343)
}
